contest-generator/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
python-dotenv
requests
pydantic
schedule
